Angular now gets compiled by tsc, everything else is compiled by babel-typescript TODO: need to migrate app/angular/server to TS in order to work with tsc

+++ scripts/prepare.js
@@ -2,9 +2,10 @@
 const path = require('path');
 const shell = require('shelljs');
 const chalk = require('chalk');
+const fs = require('fs');
 const log = require('npmlog');
 const { babelify } = require('./compile-js');
-const { generateTSDefinitionFiles } = require('./compile-ts');
+const { tscfy } = require('./compile-ts');
 
 function getPackageJson() {
   const modulePath = path.resolve('./');
@@ -21,11 +22,11 @@ function cleanup() {
   // add .ts filtering to babel args and remove after babel - 7 is adopted
   // --copy-files option doesn't work with --ignore
   // https://github.com/babel/babel/issues/5404
-
-  const files = shell.find('dist').filter(tsFile => tsFile.match(/\.(ts|tsx)$/));
-
-  if (files.length) {
-    shell.rm(files);
+  if (fs.existsSync(path.join(process.cwd(), 'dist'))) {
+    const files = shell.find('dist').filter(tsFile => tsFile.match(/\.(ts|tsx)$/));
+    if (files.length) {
+      shell.rm(files);
+    }
   }
 }
 
@@ -38,13 +39,8 @@ function logError(type, packageJson) {
 const packageJson = getPackageJson();
 
 removeDist();
-// if (packageJson && packageJson.types && packageJson.types.indexOf('d.ts') !== -1) {
-//   tscfy({ errorCallback: () => logError('ts', packageJson) });
-// } else {
 babelify({ errorCallback: () => logError('js', packageJson) });
 cleanup();
-generateTSDefinitionFiles({ errorCallback: () => logError('ts', packageJson) });
-
-// }
+tscfy({ errorCallback: () => logError('ts', packageJson) });
 
 console.log(chalk.gray(`Built: ${chalk.bold(`${packageJson.name}@${packageJson.version}`)}`));
